chore : 에러코드 수정
