+++ src/main/java/co/kr/allpick/domain/member/service/AuthService.java
@@ -1,46 +1,10 @@
 package co.kr.allpick.domain.member.service;
 
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
 import co.kr.allpick.domain.member.dto.AuthResponseDto;
 import co.kr.allpick.domain.member.dto.LoginRequestDto;
 import co.kr.allpick.domain.member.dto.SignupRequestDto;
-import co.kr.allpick.domain.member.entity.Member;
-import co.kr.allpick.domain.member.repository.MemberRepository;
-import co.kr.allpick.global.util.JwtUtil;
-import lombok.RequiredArgsConstructor;
 
-@Service
-@RequiredArgsConstructor
-public class AuthService {
-
-    private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtUtil jwtUtil;
-
-    public void signup(SignupRequestDto dto) {
-        if (memberRepository.existsByEmail(dto.getEmail())) {
-            throw new RuntimeException("이미 사용 중인 이메일입니다.");
-        }
-        Member member = Member.create(
-            dto.getEmail(),
-            passwordEncoder.encode(dto.getPassword()),
-            dto.getName(),
-            dto.getPhone()
-        );
-        memberRepository.save(member);
-    }
-
-    public AuthResponseDto login(LoginRequestDto dto) {
-        Member member = memberRepository.findByEmail(dto.getEmail())
-                .orElseThrow(() -> new RuntimeException("이메일 또는 비밀번호가 틀렸습니다"));
-
-        if (!passwordEncoder.matches(dto.getPassword(), member.getPassword())) {
-            throw new RuntimeException("이메일 또는 비밀번호가 틀렸습니다");
-        }
-
-        String token = jwtUtil.generateToken(member.getEmail());
-        return new AuthResponseDto(token, member.getEmail(), member.getName());
-    }
+public interface AuthService {
+    void signup(SignupRequestDto dto);
+    AuthResponseDto login(LoginRequestDto dto);
 }
